We got back a vehicle_id of 179 billion.  Use a long.

+++ TeslaLib/TeslaVehicle.cs
@@ -42,7 +42,7 @@ namespace TeslaLib
         */
 
         [JsonProperty(PropertyName = "vehicle_id")]
-        public int VehicleId { get; set; }
+        public long VehicleId { get; set; }
 
         [JsonProperty(PropertyName = "vin")]
         public string Vin { get; set; }
